requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
confection==0.0.4
crcmod==1.7
cymem==2.0.7
dill==0.3.1.1
dnspython==2.3.0
docopt==0.6.2
fastavro==1.7.4
fasteners==0.18
filelock==3.12.0
fsspec==2023.6.0
grpcio==1.54.0
hdfs==2.7.0
httplib2==0.21.0
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.24.3
objsize==0.6.1
ohnlp-backbone-xlang-python==1.0.10
orjson==3.8.12
packaging==23.1
pathy==0.10.1
phonenumbers==8.13.13
preshed==3.0.8
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
proto-plus==1.22.2
protobuf==4.22.5
py4j==0.10.9.7
pyarrow==11.0.0
pycryptodome==3.18.0
pydantic==1.10.9
pydot==1.4.2
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-file==1.5.1
safetensors==0.3.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
tldextract==3.4.4
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
wasabi==1.1.2
zstandard==0.21.0
